+++ spec/graphql/queries/bars_query_spec.rb
@@ -2,7 +2,7 @@
 
 require "rails_helper"
 
-describe "BarsQuery", type: :query do
+describe "BarsQuery" do
   let(:query) do
     <<~GRAPHQL
       {
